# openAPI.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
paths:
...
  /api/v1/questions:
    post:
      tags:
        - question
      operationId: createQuestion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionInputDto'
        required: true
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponseQuestionOutputDto'
  /api/v1/questions/{id}:
    get:
      tags:
        - question
      operationId: findByIdQuestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponseQuestionOutputDto'
  /api/v1/questions/quiz/{quizId}:
    get:
      tags:
        - question
      operationId: getQuestionsByQuiz
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponseListQuestionOutputDto'
  /api/v1/quiz:
    post:
      tags:
        - quiz
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizInputDto'
        required: true
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponseQuizOutputDto'
  /api/v1/quiz/{id}:
    get:
      tags:
        - quiz
      operationId: findByIdQuiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseResponseQuizOutputDto'
...
components:
  schemas:
    BaseResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/Status'
        data:
          type: object
    Status:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    ChoiceInputDto:
      required:
        - choiceText
        - correct
        - order
      type: object
      properties:
        choiceText:
          type: string
        correct:
          type: boolean
        order:
          type: integer
          format: int64
    QuestionInputDto:
      required:
        - choices
        - multipleChoice
        - order
        - questionText
        - quizId
      type: object
      properties:
        questionText:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/ChoiceInputDto'
        order:
          type: integer
          format: int64
        quizId:
          type: integer
          format: int64
        multipleChoice:
          type: boolean
...
    BaseResponseQuestionOutputDto:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/Status'
        data:
          $ref: '#/components/schemas/QuestionOutputDto'
...
    ChoiceOutputDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        choiceText:
          type: string
        order:
          type: integer
          format: int64
...
    QuestionOutputDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        questionText:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/ChoiceOutputDto'
        order:
          type: integer
          format: int64
        multipleChoice:
          type: boolean
...
    BaseResponseListQuestionOutputDto:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/Status'
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOutputDto'
    BaseResponseQuizOutputDto:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/Status'
        data:
          $ref: '#/components/schemas/QuizOutputDto'
    QuizInputDto:
      required:
        - configId
        - name
        - authorName
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        configId:
          type: integer
          format: int64
        authorName:
          type: string
    QuizOutputDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        configId:
          type: integer
          format: int64
        authorName:
          type: string
...